Updated datatypes of transaction and wallet model to BigInt (#53)

+++ app/models/transaction.py
@@ -1,7 +1,7 @@
 import datetime
 from typing import Union, List
 
-from sqlalchemy import Column, Integer, String, DateTime, or_, desc
+from sqlalchemy import Column, Integer, String, DateTime, BigInteger, or_, desc
 from sqlalchemy.orm import Query
 
 from app import wrapper
@@ -10,13 +10,13 @@ from app import wrapper
 class Transaction(wrapper.Base):
     __tablename__: str = "currency_transaction"
 
-    id: Union[Column, int] = Column(Integer, primary_key=True, autoincrement=True, unique=True)
+    id: Union[Column, int] = Column(BigInteger, primary_key=True, autoincrement=True, unique=True)
     time_stamp: Union[Column, datetime.datetime] = Column(DateTime, nullable=False)
     source_uuid: Union[Column, str] = Column(String(36))
-    send_amount: Union[Column, int] = Column(Integer, nullable=False, default=0)
+    send_amount: Union[Column, int] = Column(BigInteger, nullable=False, default=0)
     destination_uuid: Union[Column, str] = Column(String(36))
     usage: Union[Column, str] = Column(String(255), default="")
-    origin: Union[Column, Integer] = Column(Integer)
+    origin: Union[Column, Integer] = Column(BigInteger)
 
     @property
     def serialize(self) -> dict:
